requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.1
Pillow==10.2.0
celery==5.3.6
redis==5.0.3
django-cors-headers==4.3.1
drf-nested-routers==0.93.5
django-filter==23.5
